- Überprüfen ob "data.json" bereits existiert - bei post wird nun der username mitgegeben

+++ app.py
@@ -1,14 +1,15 @@
 import json
 import api_utils
-import json
+import os
 
 api = api_utils.API()
 filename = "data.json"
 
-# prepare data.json
-with open(filename, "w") as f:
-    json.dump([], f)
-    f.close()
+# prepare data.json if it doesn't exist
+if not os.path.exists(filename):
+    with open(filename, "w") as f:
+        json.dump([], f)
+        f.close()
 
 @api.GET("/api/")
 def get(request):
@@ -19,19 +20,19 @@ def get(request):
     return entryList
 
 @api.POST("/api/")
-def post(request, text:str):
-    # get current list and add new element
+def post(request, text:str, name:str):
+    # get current list and add new element with username
     with open(filename) as f:
         entryList = json.load(f)
         f.close()
-    entryList.append(text)
+    entryList.append(f"{name}: {text}")
 
     # dump new list into data.json
     with open("data.json", "w") as f:
         json.dump(entryList, f, indent=4)
         f.close()
 
-    return text
+    return None
 
 # clear list
 @api.DELETE("/api/")
